refactor: types folder
