+++ src/types/config/index.ts
@@ -1,0 +1,2 @@
+export { type DatePickerConfigType } from './DatePickerConfigType';
+export { type ThemeType } from './ThemeType';
