+++ src/types/day/index.ts
@@ -1,0 +1,2 @@
+export { type DayType } from './DayType';
+export { type DayWithTodoControls } from './DayWithTodoControls';
