+++ src/types/hooks/index.ts
@@ -1,0 +1,1 @@
+export { type UseCalendarOptionsType } from './UseCalendarOptionsType';
