+++ src/types/index.ts
@@ -1,6 +1,4 @@
-export { type DatePickerConfigType } from './DatePickerConfigType';
-export { type DayType } from './DayType';
-export { type DayWithTodoControls } from './DayWithTodoControls';
-export { type ThemeType } from './ThemeType';
-export { type TodoType } from './TodoType';
-export { type UseCalendarOptionsType } from './UseCalendarOptionsType';
+export * from './config';
+export * from './day';
+export * from './hooks';
+export * from './todos';
