+++ src/types/todos/index.ts
@@ -1,0 +1,1 @@
+export { type TodoType } from './TodoType';
